mpoove search page reducers
