+++ ui/pages/Search/reducers.js
@@ -8,8 +8,9 @@ import {
   REQUEST_SEARCHED_VARIANTS,
   RECEIVE_SEARCHED_VARIANTS,
 } from 'redux/utils/reducerUtils'
-import { loadingReducer, createSingleObjectReducer } from 'redux/utils/reducerFactories'
+import { loadingReducer, createObjectsByIdReducer, createSingleObjectReducer } from 'redux/utils/reducerFactories'
 import { HttpRequestHelper, getUrlQueryString } from 'shared/utils/httpRequestHelper'
+import { SORT_BY_XPOS } from 'shared/utils/constants'
 
 // action creators and reducers in one file as suggested by https://github.com/erikras/ducks-modular-redux
 
@@ -18,6 +19,9 @@ const REQUEST_SEARCH_CONTEXT = 'REQUEST_SEARCH_CONTEXT'
 const RECEIVE_SEARCH_CONTEXT = 'RECEIVE_SEARCH_CONTEXT'
 const REQUEST_MULTI_PROJECT_SEARCH_CONTEXT = 'REQUEST_MULTI_PROJECT_SEARCH_CONTEXT'
 const RECEIVE_MULTI_PROJECT_SEARCH_CONTEXT = 'RECEIVE_MULTI_PROJECT_SEARCH_CONTEXT'
+const REQUEST_SEARCH_GENE_BREAKDOWN = 'REQUEST_SEARCH_GENE_BREAKDOWN'
+const RECEIVE_SEARCH_GENE_BREAKDOWN = 'RECEIVE_SEARCH_GENE_BREAKDOWN'
+const UPDATE_SEARCHED_VARIANT_DISPLAY = 'UPDATE_SEARCHED_VARIANT_DISPLAY'
 
 // actions
 
@@ -139,6 +143,77 @@ export const loadSavedSearches = () => (dispatch, getState) => {
   }
 }
 
+export const updateSearchSort = updates => (dispatch) => {
+  dispatch({ type: UPDATE_SEARCHED_VARIANT_DISPLAY, updates })
+}
+
+export const loadSearchedVariants = (
+  { searchHash }, { displayUpdates, queryParams, updateQueryParams },
+) => (dispatch, getState) => {
+  const state = getState()
+  if (state.searchedVariantsLoading.isLoading) {
+    return
+  }
+
+  dispatch({ type: REQUEST_SEARCHED_VARIANTS })
+
+  let { sort, page } = displayUpdates || queryParams
+  if (!page) {
+    page = 1
+  }
+  if (!sort) {
+    sort = state.variantSearchDisplay.sort || SORT_BY_XPOS
+  }
+  const urlQueryParams = { sort: sort.toLowerCase(), page }
+
+  // Update search table state and query params
+  dispatch({ type: UPDATE_SEARCHED_VARIANT_DISPLAY, updates: { sort: sort.toUpperCase(), page } })
+  updateQueryParams(urlQueryParams)
+
+  const apiQueryParams = { ...urlQueryParams, loadFamilyContext: true, loadProjectTagTypes: true }
+  const search = state.searchesByHash[searchHash]
+  if (search && search.projectFamilies && search.projectFamilies.length > 0) {
+    apiQueryParams.loadProjectTagTypes = search.projectFamilies.some(
+      ({ projectGuid }) => !state.projectsByGuid[projectGuid]?.variantTagTypes,
+    )
+    apiQueryParams.loadFamilyContext = search.projectFamilies.some(
+      ({ familyGuids }) => !familyGuids || familyGuids.some(
+        familyGuid => !state.familiesByGuid[familyGuid]?.detailsLoaded,
+      ),
+    )
+  }
+
+  const url = `/api/search/${searchHash}?${getUrlQueryString(apiQueryParams)}`
+
+  // Fetch variants
+  new HttpRequestHelper(url,
+    (responseJson) => {
+      dispatch({ type: RECEIVE_DATA, updatesById: responseJson })
+      dispatch({ type: RECEIVE_SEARCHED_VARIANTS, newValue: responseJson.searchedVariants })
+      dispatch({ type: RECEIVE_SAVED_SEARCHES, updatesById: { searchesByHash: { [searchHash]: responseJson.search } } })
+    },
+    (e) => {
+      dispatch({ type: RECEIVE_SEARCHED_VARIANTS, error: e.message, newValue: [] })
+    }).post(search)
+}
+
+export const unloadSearchResults = () => dispatch => dispatch({ type: RECEIVE_SEARCHED_VARIANTS, newValue: [] })
+
+export const loadGeneBreakdown = searchHash => (dispatch, getState) => {
+  if (!getState().searchGeneBreakdown[searchHash]) {
+    dispatch({ type: REQUEST_SEARCH_GENE_BREAKDOWN })
+
+    new HttpRequestHelper(`/api/search/${searchHash}/gene_breakdown`,
+      (responseJson) => {
+        dispatch({ type: RECEIVE_DATA, updatesById: responseJson })
+        dispatch({ type: RECEIVE_SEARCH_GENE_BREAKDOWN, updatesById: responseJson })
+      },
+      (e) => {
+        dispatch({ type: RECEIVE_SEARCH_GENE_BREAKDOWN, error: e.message, updatesById: {} })
+      }).get()
+  }
+}
+
 // reducers
 
 export const reducers = {
@@ -147,6 +222,13 @@ export const reducers = {
     REQUEST_MULTI_PROJECT_SEARCH_CONTEXT, RECEIVE_MULTI_PROJECT_SEARCH_CONTEXT,
   ),
   flattenCompoundHet: createSingleObjectReducer(UPDATE_COMPOUND_HET_DISPLAY),
+  searchGeneBreakdown: createObjectsByIdReducer(RECEIVE_SEARCH_GENE_BREAKDOWN, 'searchGeneBreakdown'),
+  searchGeneBreakdownLoading: loadingReducer(REQUEST_SEARCH_GENE_BREAKDOWN, RECEIVE_SEARCH_GENE_BREAKDOWN),
+  variantSearchDisplay: createSingleObjectReducer(UPDATE_SEARCHED_VARIANT_DISPLAY, {
+    sort: SORT_BY_XPOS,
+    page: 1,
+    recordsPerPage: 100,
+  }, false),
 }
 
 const rootReducer = combineReducers(reducers)
